Rename article `title` and `content` to `headline` and `subhead`
These names probably originated with the early articles we scraped directly from AP's website, which were stored with different field names than are commonly used in journalism. Journalists sometimes call these the `hed` and `subhed`/`dek`, so I've adapted those names a bit to be easier for us to use while staying as true to the source domain as I can.

This differentiates the `subhead` (formerly `content`) from the actual text of the article, which we may soon start fetching from AP with additional API calls.