Create `candidate_articles` table and add store/fetch methods
This will allow us to explicitly record what we considered to be part of the candidate pool each day rather than inferring what was part of the pool from the date. I think this will help us avoid including the same article multiple days in a row, and also make it easier to do offline evaluation down the road.