Store impression ids when logging newsletters
For many of our tables, the database will auto-assign an ID value if it's missing rather than raising an error. We also have the `Impression` domain model set up to auto-assign UUIDs if they're missing. In general, these are both good behaviors that improve the flexibility and resilience of our system. However, it can also lead to a corner case where we're auto-assigning different IDs to the same impression if the impression ids attached to Python objects aren't getting stored to the database.

That's what happened with click tracking this morning: the impression ids that were getting embedded in the tracking links were the ones attached to the domain objects, which weren't getting saved to the database and therefore didn't match the database auto-assigned ids of rows in the `impressions` table. That caused a foreign-key constraint violation, even though the relevant impression rows do exist in the database (under a different id).